pyproject.toml
[tool.poetry]
name = "freetrade"
version = "0.1.0"
description = ""
authors = ["Hong"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
beautifulsoup4 = "^4.9.3"
requests-html = "^0.10.0"
gspread = "^3.7.0"
oauth2client = "^4.1.3"
google-auth-oauthlib = "^0.4.4"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
